Django/Django_Old/disa-py/disa/requirements.txt
Django==1.9.5
MySQL-python==1.2.5
bootstrap-admin==0.3.6
django-admin-bootstrapped==2.5.4
django-daterange-filter==1.1.1
django-extensions==1.6.1
gunicorn==19.4.5
git+https://github.com/darklow/django-suit.git@bs3
#django-suit==0.3a3
django-import-export==0.4.3
django-crontab==0.7.1
sqlparse==0.1.19
djangorestframework==3.3.3
django-compressor==2.0
#Markdown==2.6.6
#django-filter==0.13.0
ipython==4.1.2
openpyxl==2.3.5
django-role-permissions==1.0
python-social-auth==0.2.19
Pillow==3.3.0
